+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `splitLines` string utility function
 - `formatSwissIbanNumber` swiss standard function
+- `trimStart`, `trimEnd` and `trim` string type utility functions
 
 ### Changed